Moved checkToolExists to lib + moved tests.
