+++ lib/errcheck.go
@@ -1,6 +1,10 @@
 package lib
 
-import "os"
+import (
+	"fmt"
+	"os"
+	"os/exec"
+)
 
 // Check takes in an error and verifies if it is nil.
 // If the error is not nil, it will terminate the program.
@@ -31,4 +35,29 @@ func DirExists(dirname string) bool {
 		return false
 	}
 	return info.IsDir()
+}
+
+// CheckToolExists uses the 'which' command to find a specific tool.
+// It then parses the output of the command, and checks if
+// 'which' found the toolname in the path or not.
+// * tested
+func CheckToolExists(toolName string) bool {
+	if toolName == "" {
+		fmt.Println("Cannot pass in a blank toolname")
+		return false
+	}
+	if toolName[0] == '-' {
+		fmt.Println("Cannot pass in an option as a toolname")
+		return false
+	}
+	cmd := exec.Command("which", toolName)
+	_, err := cmd.Output()
+	// When 'which' cannot find the tool, it exits with status 1 (error)
+	if err != nil {
+		fmt.Println(err.Error())
+		fmt.Printf("PROBLEM: %s not found in path\n", toolName)
+		return false
+	}
+	fmt.Printf("%s found in path\n", toolName)
+	return true
 }
