+++ reactnative/addReactNative.go
@@ -11,36 +11,11 @@ import (
 	"os/exec"
 )
 
-// checkToolExists uses the 'which' command to find a specific tool.
-// It then parses the output of the command, and checks if
-// 'which' found the toolname in the path or not.
-// * tested
-func checkToolExists(toolName string) bool {
-	if toolName == "" {
-		fmt.Println("Cannot pass in a blank toolname")
-		return false
-	}
-	if toolName[0] == '-' {
-		fmt.Println("Cannot pass in an option as a toolname")
-		return false
-	}
-	cmd := exec.Command("which", toolName)
-	_, err := cmd.Output()
-	// When 'which' cannot find the tool, it exits with status 1 (error)
-	if err != nil {
-		fmt.Println(err.Error())
-		fmt.Printf("PROBLEM: %s not found in path\n", toolName)
-		return false
-	}
-	fmt.Printf("%s found in path\n", toolName)
-	return true
-}
-
 // installExpo checks first that npm is installed.
 // If it is, then it uses npm to globally install expo
 // * tested
 func installExpo() error {
-	if !checkToolExists("npm") {
+	if !lib.CheckToolExists("npm") {
 		return errors.New("prerequisite tool 'npm' is not installed")
 	}
 	cmd := exec.Command("npm", "i", "-g", "expo-cli")
@@ -68,7 +43,7 @@ func confirmInstall(stdin io.Reader) bool {
 // Running 'expo init' with the project name will create a new project.
 // * tested
 func AddReactNativeProj(projName string, wsPath string) {
-	expoInstalled := checkToolExists("expo")
+	expoInstalled := lib.CheckToolExists("expo")
 	ws, err := ioutil.ReadFile(wsPath)
 	lib.Check(err)
 	err = os.MkdirAll(string(ws), os.ModePerm)
